Back/requirements.txt
asgiref==3.3.4
backcall==0.2.0
decorator==5.0.9
dj-database-url==0.5.0
Django==3.2.3
django-cors-headers==3.7.0
django-heroku==0.3.1
djangorestframework==3.12.4
Faker==8.10.1
gunicorn==20.1.0
ipdb==0.13.8
ipython==7.23.1
ipython-genutils==0.2.0
jedi==0.18.0
matplotlib-inline==0.1.2
parso==0.8.2
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.18
psycopg2==2.9.1
ptyprocess==0.7.0
Pygments==2.9.0
python-dateutil==2.8.2
pytz==2021.1
six==1.16.0
sqlparse==0.4.1
text-unidecode==1.3
toml==0.10.2
traitlets==5.0.5
wcwidth==0.2.5
whitenoise==5.3.0
